+++ src/hooks/macroActivity/loadActivity.js
@@ -1,9 +1,9 @@
-import MacroActivity from "./macroActivity.js";
+import MacroActivity from "./MacroActivity.js";
 
 
 export default function addActivitiesHooks() {
-  CONFIG.DND5E.activityTypes["macro"] = {
-    name: "Macro",
+  CONFIG.DND5E.activityTypes["ddbmacro"] = {
+    // hidden: true,
     documentClass: MacroActivity,
   };
 }